+++ src/store/index.ts
@@ -1,5 +1,6 @@
 export * from "./backup";
 export * from "./database";
 export * from "./errors";
+export * from "./import-export";
 export * from "./schema";
 export * from "./syx-codec";